conflict solves and some modifications on KNN(not working yet)

+++ src/main/java/org/graphast/query/route/shortestpath/dijkstra/DijkstraGeneric.java
@@ -122,7 +122,7 @@ public class DijkstraGeneric {
 	    HashMap<Long, Integer> shortestDistances = new HashMap<Long, Integer>();
 	    
 	    shortestDistances.put(v, 0);
-	    QueueEntry e = new QueueEntry(v, (short) 0);
+	    QueueEntry e = new QueueEntry(v, 0);
         unsettledNodes.add(e);
         
         while ((e = unsettledNodes.poll()) != null){
